rename: type 관련 파일들 파일명.type.ts로 변경
