+++ src/types/card.type.ts
