+++ src/types/props.type.ts
